# greenon-server/src/swagger/swagger.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Greenon API
paths:
  /login:
    post:
      tags:
        - Authentication
      summary: 로그인
      parameters:
        - in: query
          name: email
          schema:
            type: string
          required: true
        - in: query
          name: password
          schema:
            type: string
          required: true
      responses:
        200:
          description: Success
        400:
          description: Invalid input
  /register:
    post:
      tags:
        - Authentication
      summary: 회원가입
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: email
          required: true
          schema:
            type: string
        - in: query
          name: password
          required: true
          schema:
            type: string
      responses:
        200:
          description: Success
  /kakao-login:
    post:
      summary: 카카오 로그인
      tags:
        - Authentication
      responses:
        200:
          description: Success
  /apple-login:
    post:
      summary: 애플 로그인
      tags:
        - Authentication
      responses:
        200:
          description: Success
  /sms:
    post:
      summary: SMS 인증
      tags:
        - Authentication
      responses:
        200:
          description: Success
  /users:
    get:
      summary: 유저 목록 조회
      tags:
        - User
      parameters:
        - $ref: parameters.yaml#/token
      responses:
        200:
          description: Success
    put:
      summary: 유저 정보 수정
      tags:
        - User
      parameters:
        - $ref: parameters.yaml#/token
      responses:
        200:
          description: Success
    delete:
      tags:
        - User
      summary: 유저 삭제
      parameters:
        - $ref: parameters.yaml#/token
      responses:
        200:
          description: Success
  /devices:
    get:
      tags:
        - Device
      summary: 디바이스 목록 조회
      parameters:
        - $ref: parameters.yaml#/token
      responses:
        200:
          description: Success
    post:
      tags:
        - Device
      summary: 디바이스 추가
      parameters:
        - $ref: parameters.yaml#/token
      responses:
        200:
          description: Success

  /devices/:pk:
    get:
      tags:
        - Device Data
      summary: 디바이스 데이터 조회
      parameters:
        - $ref: parameters.yaml#/token
      responses:
        200:
          description: Success